add restriction with variable with width != 32

+++ books/projects/rac/tests/yaml_test/control_flow/for-loop-invalid.cpp
@@ -4,8 +4,8 @@
 
 typedef ac_int<8, false> u8;
 
-u8 adder(u8 a, u8 b) {
-  u8 res = 0;
+int adder(u8 a, u8 b) {
+  int res = 0;
   for (res = b; a != 0; a--)
     res++;
